backend/requirements.txt
Flask==3.1.0
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.1.0
